+++ src/HetrickCV.cpp
@@ -48,6 +48,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelPhasorRanger);
 	p->addModel(modelPhasorReset);
 	p->addModel(modelPhasorShape);
+	p->addModel(modelPhasorSubstepShape);
 	p->addModel(modelPhasorToLFO);
 	p->addModel(modelProbability);
 	p->addModel(modelRandomGates);