+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - Add Phasor->Gates 32 module.
     - Add Phasor->Gates 64 module.
     - Add Phasor->LFO module.
+    - Add Phasor Reset module.
 - Add Gate Delay module.
 - Add Thomas chaotic attractor to 1-Op Chaos (https://sprott.physics.wisc.edu/chaos/symmetry.htm).
 - Make Mode knobs on 1-Op Chaos, 2-Op Chaos, Chaotic Attactors, and Clocked Noise snap to values and print the current mode.